requirements.txt
black==22.3.0
flake8==4.0.1

# upload and such, get rid of these!
azure-storage-blob==12.16.0
azure-identity==1.13.0
azure-mgmt-cognitiveservices==13.4.0
azure-mgmt-resource==23.0.1
azure-mgmt-network==23.1.0
azure-mgmt-compute==30.0.0
azure-mgmt-storage==21.0.0

# ssh dependancies
paramiko==3.2.0

#TODO test dependancies
pytest==6.2.4

#TODO packageing dependancies
wheel==0.40.0
twine==4.0.2




